fix(gui): silence deprecated enum-enum bitwise OR in CReconnectDialog
wxALIGN_CENTER (wxAlignment) OR'd with wxLEFT/wxRIGHT/wxBOTTOM
(wxDirection) mixes two different enum types, which C++20 deprecates
and GCC/Clang warn on (-Wdeprecated-enum-enum-conversion). Cast the
first operand to int, matching the existing pattern used elsewhere in
this file's sibling sizer calls (FirstRunWizard.cpp, amule.cpp).

+++ src/amule-remote-gui.cpp
@@ -148,7 +148,7 @@ public:
 		top->Add(m_label, 0, wxALL, 15);
 		top->Add(new wxButton(this, wxID_CANCEL, _("Abort and exit")),
 			0,
-			wxALIGN_CENTER | wxLEFT | wxRIGHT | wxBOTTOM,
+			static_cast<int>(wxALIGN_CENTER) | wxLEFT | wxRIGHT | wxBOTTOM,
 			15);
 		SetAttempt(1);
 		SetSizerAndFit(top);
